+++ src/components/header/index.test.tsx
@@ -1,0 +1,10 @@
+import { render, screen } from '@testing-library/react'
+import { Header } from '.'
+
+describe('Header', () => {
+  it('should render correctly', () => {
+    render(<Header />)
+
+    expect(screen.getByRole('banner')).toBeInTheDocument()
+  })
+})