+++ src/components/header/theme-menu/index.tsx
@@ -2,60 +2,35 @@ import { Fragment, useEffect } from 'react'
 import { Menu, Transition } from '@headlessui/react'
 import { PiPaintBrushBroadBold } from 'react-icons/pi'
 
-import { Theme, useThemeStore } from '@/stores/theme-store'
+import useQueryState from '~/shared/query/query-state'
+import { Theme, useThemeStore } from '~/stores/theme-store'
+import { themeList } from './theme-list'
 import { themeButton, triggerButton } from './styles'
 
 export function ThemeMenu() {
   const setTheme = useThemeStore(set => set.setTheme)
   const currentTheme = useThemeStore(set => set.theme)
+  const [queryTheme, setQueryTheme] = useQueryState('theme')
 
   function handleTheme(newTheme: Theme) {
     setTheme(newTheme)
+    setQueryTheme(newTheme)
     localStorage.setItem('theme', newTheme)
   }
 
   useEffect(() => {
     if (typeof window !== 'undefined') {
       const themeOnStorage = localStorage.getItem('theme') as Theme
-      if (themeOnStorage) setTheme(themeOnStorage)
-      else localStorage.setItem('theme', currentTheme)
+      if (queryTheme) {
+        setTheme(queryTheme as Theme)
+      } else if (themeOnStorage) {
+        setTheme(themeOnStorage)
+        setQueryTheme(themeOnStorage)
+      } else {
+        localStorage.setItem('theme', currentTheme)
+      }
     }
   }, [])
-
-  interface ThemeAttrs {
-    title: string
-    id: Theme
-  }
-  const themeList: ThemeAttrs[] = [
-    {
-      title: 'Light',
-      id: 'light'
-    },
-    {
-      title: 'Dark',
-      id: 'dark'
-    },
-    {
-      title: 'Transition',
-      id: 'transition'
-    },
-    {
-      title: 'Blue Room',
-      id: 'blue-room'
-    },
-    {
-      title: 'Train',
-      id: 'train'
-    },
-    {
-      title: 'Waterfall',
-      id: 'waterfall'
-    },
-    {
-      title: 'Camping Fire',
-      id: 'camping-fire'
-    }
-  ]
 
   return (
     <div className="z-40">
@@ -80,6 +55,7 @@ export function ThemeMenu() {
         >
           <Menu.Items
             as="div"
+            data-testid="theme-menu-container"
             className="snap absolute left-4 right-4 mt-2 flex max-h-[85vh] origin-top-right snap-y snap-mandatory scroll-pt-4 flex-col gap-4 overflow-y-auto rounded-2xl p-4 shadow-2xl backdrop-blur-md scrollbar-hide focus:outline-none md:left-auto md:right-0"
           >
             {themeList.map(theme => {