+++ src/sounds.tsx
@@ -22,7 +22,7 @@ import {
   Waves,
   WhiteNoise,
   Wind
-} from '@/components/sound-icons'
+} from '~/components/sound-icons'
 
 export const sounds = [
   {
@@ -30,7 +30,7 @@ export const sounds = [
     title: 'Rain',
     icon: Rain,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/rain.ogg',
+      url: 'https://cdn.noisekun.com/sounds/rain.ogg',
       type: 'audio/ogg'
     }
   },
@@ -39,7 +39,7 @@ export const sounds = [
     title: 'Storm',
     icon: Storm,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/storm.ogg',
+      url: 'https://cdn.noisekun.com/sounds/storm.ogg',
       type: 'audio/ogg'
     }
   },
@@ -48,7 +48,7 @@ export const sounds = [
     title: 'Drops',
     icon: Drops,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/drops.ogg',
+      url: 'https://cdn.noisekun.com/sounds/drops.ogg',
       type: 'audio/ogg'
     }
   },
@@ -57,7 +57,7 @@ export const sounds = [
     title: 'Wind',
     icon: Wind,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/wind.ogg',
+      url: 'https://cdn.noisekun.com/sounds/wind.ogg',
       type: 'audio/ogg'
     }
   },
@@ -66,7 +66,7 @@ export const sounds = [
     title: 'Waves',
     icon: Waves,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/waves.ogg',
+      url: 'https://cdn.noisekun.com/sounds/waves.ogg',
       type: 'audio/ogg'
     }
   },
@@ -75,7 +75,7 @@ export const sounds = [
     title: 'Underwater',
     icon: Underwater,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/underwater.ogg',
+      url: 'https://cdn.noisekun.com/sounds/underwater.ogg',
       type: 'audio/ogg'
     }
   },
@@ -84,7 +84,7 @@ export const sounds = [
     title: 'Stream Water',
     icon: StreamWater,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/stream-water.ogg',
+      url: 'https://cdn.noisekun.com/sounds/stream-water.ogg',
       type: 'audio/ogg'
     }
   },
@@ -93,7 +93,7 @@ export const sounds = [
     title: 'Waterfall',
     icon: Waterfall,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/waterfall.ogg',
+      url: 'https://cdn.noisekun.com/sounds/waterfall.ogg',
       type: 'audio/ogg'
     }
   },
@@ -102,7 +102,7 @@ export const sounds = [
     title: 'Birds on Tree',
     icon: BirdsTree,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/birds-tree.ogg',
+      url: 'https://cdn.noisekun.com/sounds/birds-tree.ogg',
       type: 'audio/ogg'
     }
   },
@@ -111,7 +111,7 @@ export const sounds = [
     title: 'Leaves',
     icon: Leaves,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/leaves.ogg',
+      url: 'https://cdn.noisekun.com/sounds/leaves.ogg',
       type: 'audio/ogg'
     }
   },
@@ -120,7 +120,7 @@ export const sounds = [
     title: 'Bonfire',
     icon: Fire,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/fire.ogg',
+      url: 'https://cdn.noisekun.com/sounds/fire.ogg',
       type: 'audio/ogg'
     }
   },
@@ -129,7 +129,7 @@ export const sounds = [
     title: 'Cave Sounds',
     icon: Cave,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/cave-drops.ogg',
+      url: 'https://cdn.noisekun.com/sounds/cave-drops.ogg',
       type: 'audio/ogg'
     }
   },
@@ -138,7 +138,7 @@ export const sounds = [
     title: 'Night',
     icon: Night,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/night.ogg',
+      url: 'https://cdn.noisekun.com/sounds/night.ogg',
       type: 'audio/ogg'
     }
   },
@@ -147,7 +147,7 @@ export const sounds = [
     title: 'Coffee shop',
     icon: Coffee,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/coffee.ogg',
+      url: 'https://cdn.noisekun.com/sounds/coffee.ogg',
       type: 'audio/ogg'
     }
   },
@@ -156,7 +156,7 @@ export const sounds = [
     title: 'Train',
     icon: Train,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/train.ogg',
+      url: 'https://cdn.noisekun.com/sounds/train.ogg',
       type: 'audio/ogg'
     }
   },
@@ -165,7 +165,7 @@ export const sounds = [
     title: 'Airplane',
     icon: AirPlane,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/air-plane.ogg',
+      url: 'https://cdn.noisekun.com/sounds/air-plane.ogg',
       type: 'audio/ogg'
     }
   },
@@ -174,7 +174,7 @@ export const sounds = [
     title: 'Washing machine',
     icon: WashingMachine,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/washing-machine.ogg',
+      url: 'https://cdn.noisekun.com/sounds/washing-machine.ogg',
       type: 'audio/ogg'
     }
   },
@@ -183,7 +183,7 @@ export const sounds = [
     title: 'Playground',
     icon: Playground,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/playground.ogg',
+      url: 'https://cdn.noisekun.com/sounds/playground.ogg',
       type: 'audio/ogg'
     }
   },
@@ -192,7 +192,7 @@ export const sounds = [
     title: 'Boat',
     icon: Boat,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/boat.ogg',
+      url: 'https://cdn.noisekun.com/sounds/boat.ogg',
       type: 'audio/ogg'
     }
   },
@@ -201,7 +201,7 @@ export const sounds = [
     title: 'Rain on Tent',
     icon: RainOnTent,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/rain-on-tent.ogg',
+      url: 'https://cdn.noisekun.com/sounds/rain-on-tent.ogg',
       type: 'audio/ogg'
     }
   },
@@ -210,7 +210,7 @@ export const sounds = [
     title: 'Brown Noise',
     icon: BrownNoise,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/brown-noise.ogg',
+      url: 'https://cdn.noisekun.com/sounds/brown-noise.ogg',
       type: 'audio/ogg'
     }
   },
@@ -219,7 +219,7 @@ export const sounds = [
     title: 'White Noise',
     icon: WhiteNoise,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/white-noise.ogg',
+      url: 'https://cdn.noisekun.com/sounds/white-noise.ogg',
       type: 'audio/ogg'
     }
   },
@@ -228,7 +228,7 @@ export const sounds = [
     title: 'Pink Noise',
     icon: PinkNoise,
     file: {
-      url: 'https://cdn.noisekun.mateusf.com/sounds/pink-noise.ogg',
+      url: 'https://cdn.noisekun.com/sounds/pink-noise.ogg',
       type: 'audio/ogg'
     }
   }